requirements.txt
fastapi
uvicorn
cryptography
tinyec
pycryptodome